// src/app/core/model/card.model.ts
export enum Brand {
    UNKNOWN = 0,
    VISA = 2,
    MASTER_CARD = 3,
    BIT_COIN = 5,
    PAY_PAL = 7,
    DISCOVER = 11,
    AMEX = 13,
    JCB = 15,
    BTC = 17,
    DINERS = 19,
    STRIPE = 22,
    CHINA_UNION_PAY = 22,
    MAESTRO = 27,
    SWITCH = 29,
    CARTE_BLANCHE = 31,
    SOLO = 33,
    OTHER = 79
}